fix: fix eslint config paths and import order issues after restructuring
- Fix ee-embed-sdk .eslintrc.json path (was 4 levels up, now 3)
- Fix engine .eslintrc.json path (now sibling to api under server/)
- Auto-fix import order lint errors in shared package
- Fix test imports to use new automation/ subdirectory paths
