+++ packages/ee/embed-sdk/.eslintrc.json
@@ -1,5 +1,5 @@
 {
-  "extends": ["../../../../.eslintrc.base.json"],
+  "extends": ["../../../.eslintrc.base.json"],
   "ignorePatterns": ["!**/*"],
   "overrides": [
     {
