+++ packages/shared/src/lib/automation/agents/index.ts
@@ -1,6 +1,6 @@
 import { Static, Type } from '@sinclair/typebox'
-import { AIProviderName } from '../../management/ai-providers'
 import { DiscriminatedUnion, Nullable } from '../../core/common'
+import { AIProviderName } from '../../management/ai-providers'
 export * from './tools'
 export * from './mcp'
