+++ packages/shared/src/lib/automation/engine/engine-operation.ts
@@ -1,12 +1,12 @@
 import { Static, Type } from '@sinclair/typebox'
+import { PlatformId } from '../../management/platform'
+import { ProjectId } from '../../management/project/project'
 import { ExecutionToolStatus, PredefinedInputsStructure } from '../agents'
 import { AppConnectionValue } from '../app-connection/app-connection'
 import { ExecutionState, ExecutionType, ResumePayload } from '../flow-run/execution/execution-output'
 import { FlowRunId, RunEnvironment } from '../flow-run/flow-run'
 import { FlowVersion } from '../flows/flow-version'
 import { PiecePackage } from '../pieces'
-import { PlatformId } from '../../management/platform'
-import { ProjectId } from '../../management/project/project'
 import { ScheduleOptions } from '../trigger'
 
 export enum EngineOperationType {
