+++ packages/shared/src/lib/ee/billing/index.ts
@@ -1,7 +1,7 @@
-import { isNil, Nullable } from '../../core/common'
-import { PiecesFilterType } from '../../management/project'
-import { AiCreditsAutoTopUpState, PlanName, PlatformPlanWithOnlyLimits, PlatformUsageMetric, TeamProjectsLimit } from '../../management/platform'
 import { Static, Type } from '@sinclair/typebox'
+import { isNil, Nullable } from '../../core/common'
+import { AiCreditsAutoTopUpState, PlanName, PlatformPlanWithOnlyLimits, PlatformUsageMetric, TeamProjectsLimit } from '../../management/platform'
+import { PiecesFilterType } from '../../management/project'
 
 export const PRICE_PER_EXTRA_ACTIVE_FLOWS = 5
 export const AI_CREDITS_USAGE_THRESHOLD = 15000
