+++ packages/shared/src/lib/ee/custom-domains/index.ts
@@ -1,5 +1,5 @@
-import { BaseModelSchema } from '../../core/common/base-model'
 import { Static, Type } from '@sinclair/typebox'
+import { BaseModelSchema } from '../../core/common/base-model'
 
 export enum CustomDomainStatus {
     ACTIVE = 'ACTIVE',
