+++ packages/shared/src/lib/ee/signing-key/signing-key-model.ts
@@ -1,6 +1,6 @@
+import { Static, Type } from '@sinclair/typebox'
 import { BaseModelSchema } from '../../core/common/base-model'
 import { ApId } from '../../core/common/id-generator'
-import { Static, Type } from '@sinclair/typebox'
 
 export enum KeyAlgorithm {
     RSA = 'RSA',
